requirements.txt

pandas
flywheel-sdk
